Typoes and bug fixes.

+++ core/src/com/majalis/character/PlayerCharacter.java
@@ -968,7 +968,7 @@ public class PlayerCharacter extends AbstractCharacter {
 		}
 		if (sex.isOgreSex() && perks.get(Perk.GIANT_LOVER.toString(), 0) != 3) {
 			result += "You gained " + Perk.GIANT_LOVER.getLabel() + " (Rank " + (perks.get(Perk.GIANT_LOVER.toString(), 0) + 1) + ")!";
-			perks.put(Perk.GIANT_LOVER.toString(), perks.get(Perk.GIANT_LOVER.toString(), 0) + 1);
+			perks.put(Perk.GIANT_LOVER.toString(), ((int)perks.get(Perk.GIANT_LOVER.toString(), 0)) + 1);
 		}
 		
 		return result;
